Refactor performance profiling into its own overlay
