+++ src/main/java/com/eerussianguy/blazemap/feature/maps/Overlays.java
@@ -13,9 +13,11 @@ import static com.eerussianguy.blazemap.BlazeMap.MOD_NAME;
 
 public class Overlays {
     public static final IIngameOverlay MINIMAP = OverlayRegistry.registerOverlayTop(MOD_NAME + " Minimap", Overlays::renderMinimap);
+    public static final IIngameOverlay PROFILER = OverlayRegistry.registerOverlayTop(MOD_NAME + " Profiler", Overlays::renderProfiler);
 
     public static void reload() {
         OverlayRegistry.enableOverlay(MINIMAP, BlazeMapConfig.CLIENT.enableMinimap.get());
+        OverlayRegistry.enableOverlay(PROFILER, BlazeMapConfig.CLIENT.enableDebug.get());
     }
 
     public static void renderMinimap(ForgeIngameGui gui, PoseStack stack, float partialTicks, int width, int height) {
@@ -23,6 +25,15 @@ public class Overlays {
         stack.scale(0.5f, 0.5f, 1f);
         var buffers = MultiBufferSource.immediate(Tesselator.getInstance().getBuilder());
         MinimapRenderer.INSTANCE.draw(stack, buffers, gui, width * 2, height * 2);
+        buffers.endBatch();
+        stack.popPose();
+    }
+
+    public static void renderProfiler(ForgeIngameGui gui, PoseStack stack, float partialTicks, int width, int height) {
+        stack.pushPose();
+        stack.scale(0.5f, 0.5f, 1f);
+        var buffers = MultiBufferSource.immediate(Tesselator.getInstance().getBuilder());
+        ProfilingRenderer.INSTANCE.draw(stack, buffers, gui, width * 2, height * 2);
         buffers.endBatch();
         stack.popPose();
     }
